requirements.txt
langchain==0.1.13
langchain_community==0.0.29
langchain_core==0.1.33
langchain_openai==0.1.0
python-dotenv==1.0.1
streamlit==1.32.2
